% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_temp.R
\name{convert_temp}
\alias{convert_temp}
\title{Classify the Weather}
\usage{
convert_temp(temperature, method)
}
\arguments{
\item{temperature}{the temperature in C or F}

\item{method}{either "c2f" or "f2c" conversion methods}
}
\description{
Tells us the temperature in Celsius or Fahrenheit depending on the given method
}
\examples{
#example code
convert_temp(10,"c2f")

}
